ft_split a meio

+++ ft_memset.c
@@ -3,21 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: pesilva- <pesilva-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/09 09:41:27 by marvin            #+#    #+#             */
-/*   Updated: 2024/04/09 09:41:27 by marvin           ###   ########.fr       */
+/*   Created: 2024/04/10 13:06:44 by pesilva-          #+#    #+#             */
+/*   Updated: 2024/04/10 13:06:44 by pesilva-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 #include "libft.h"
 #include <stdio.h>
 #include <string.h>
 
-void	*ft_memset(void *ptr, int value, int num)
+void	*ft_memset(void *ptr, int value, size_t num)
 {
 	unsigned char	*char_ptr;
-	int				i;
+	size_t			i;
 
 	i = 0;
 	char_ptr = ptr;
